Make Test a weak dependency
